+++ packages/core/CHANGELOG.md
@@ -8,7 +8,7 @@
 - Feature: Add `deviceIsHorizontal` as a getter to have a more central place to check if the device is in landscape mode.
 - Feature: Add clearer documentation regarding `@masterportal/masterportalapi` related configuration parameters including examples.
 - Feature: Officially add support for WMTS layers.
-- Feature: Add reasonable default values for configuration parameters `epsg`, `backgroundImage`, `options`, `namedProjections` and `startResolution`.
+- Feature: Add reasonable default values for configuration parameters `epsg`, `options`, `namedProjections` and `startResolution`.
 - Fix: Document `rawLayerList.initializeLayerList` as an essential step when creating a client.
 - Fix: Move basic documentation of `layers` from `@polar/plugin-layer-chooser` to `@polar/core`.
 - Fix: Adjust documentation to properly describe optionality of configuration parameters.